Move non-provider stuff in AuthProvider into new file
This also resulted in changing test-utils, since the exports of the
auth-related keys have been replaced with more proper exports.

This also preemptively partially removes the "post" action from
authReducer for compatibility with #33.
