+++ src/auth/__tests__/PrivateRoute.unit.test.tsx
@@ -4,10 +4,11 @@ import { Router, Switch } from "react-router-dom";
 import { createMemoryHistory, History } from "history";
 import { render } from "@testing-library/react";
 import { act } from "react-dom/test-utils";
-import { USER_LOCALSTORAGE_KEY, testUser, testAdmin } from "../../shared/test-utils";
+import { testUser, testAdmin } from "../../shared/test-utils";
 import mockAxios from "jest-mock-axios";
 import PrivateRoute from "../PrivateRoute";
-import { Auth, AuthProvider } from "../AuthProvider";
+import { Auth, USER_LOCALSTORAGE_KEY } from "../Auth";
+import { AuthProvider } from "../AuthProvider";
 import ERRORS from "../../shared/errors";
 
 describe("Unit test for PrivateRoute", () => {
